New branch fixes several issues:
  - Handle 'show ratings' per itemtype as a modvar
  - Change db to allow rating values that align with form values (ie, 0-100) 
    rather than 0 to 9.99999
  - Adapt xaruser/display.php, xaruserapi/rate.php and xaruser/rate.php to 
    handle stored ratings correctly
  - Change xartemplates/user-display.php to show all stars (whole, half, 
    empty) accoring to style setting
  - Changed xarimages/halfstar.gif to show entire star, half full
  - Added xarML calls around star image alt text
  - Version increased from 1.2.1 to 1.3.0

Because stored ratings values have not until now been consistently treated 
as percentages, all exsiting ratings are skewed in such a way that accurately 
correcting them on upgrade is impossible as far as I can see.  The upgrade 
routine currently makes no attempt to correct them.  Input on this is 
appreciated.

+++ xaruserapi/rate.php
@@ -112,7 +112,7 @@ function ratings_userapi_rate($args)
 
         // Calculate new rating
         $newnumratings = $numratings + 1;
-        $newrating = (int)((($currating*$numratings) + $rating)/$newnumratings);
+        $newrating = (($currating*$numratings) + $rating)/$newnumratings;
 
         // Insert new rating
         $query = "UPDATE $ratingstable
